Finally started working on voice chat support.
I have created VoiceResponse which should probably just have been named 'VoiceChat' but that can be renamed later.
This will listen to the people in the joined voice chat and when they are quiet for 2 seconds, it will process the user's audio extract the speech as text using local hugging face ASR model. I'm using whisper-medium while developing and it does a very good job with the testing I have done so far. I just need to get GPU support working with the nvidea GPU.
It will then send that voice chat snippet as text to Ollama chat with the previous chat history from that voice channel as context and give a response. Right now it doesn't actually play the response, it just logs it in the message history for that voice channel.
I have only tested around with 1 person in the voice chat along with the bot, but it should be able to handle multiple peoples voices and send it to the ollama chat. There are probably more async issues I'm overlooking right now.